# cmd/wallet-doc-gen/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
        "/healthz": {
            "get": {
                "description": "系统健康检查",
                "consumes": [
                    "application/json"
                ],
                "tags": [
                    "系统服务"
                ],
                "summary": "健康检查",
                "responses": {
                    "200": {
                        "description": "{\"code\": 200, \"data\":{\"status\": \"ok\"} }",
                        "schema": {
                            "$ref": "#/definitions/server.Response"
                        }
                    }
                }
            }
        },
        "/v1/iamgeek/chains": {
            "get": {
                "description": "获取接口管理列表",
                "tags": [
                    "区块链管理"
                ],
                "summary": "获取区块链列表",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "页条数",
                        "name": "limit",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "页码",
                        "name": "offset",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "{\"code\": 200, \"message\":\"Ok\",\"data\": [...]}",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/core.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/v1.BlockChainList"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/core.Response"
                        }
                    }
                }
            },
            "post": {
                "description": "创建一个链",
                "tags": [
                    "区块链管理"
                ],
                "summary": "创建一个链",
                "parameters": [
                    {
                        "description": "链数据",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/v1.BlockChain"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "{\"code\": 200, \"message\":\"Ok\",\"data\": [...]}",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/core.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/v1.BlockChain"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/core.Response"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/core.Response"
                        }
                    }
                }
            }
        },
        "/v1/iamgeek/chains/{id}": {
            "get": {
                "description": "获取区块链详情",
                "tags": [
                    "区块链管理"
                ],
                "summary": "获取区块链详情",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "主键Id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "{\"code\": 200, message:\"ok\", \"data\": [...]}",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/core.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/v1.BlockChain"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/core.Response"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/core.Response"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/core.Response"
                        }
                    }
                }
            },
            "put": {
                "description": "更新区块链信息",
                "tags": [
                    "区块链管理"
                ],
                "summary": "更新区块链信息",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "主键id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "链数据",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/v1.BlockChain"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "{\"code\": 200, \"message\":\"Ok\", \"data\": [...]}",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/core.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/v1.BlockChain"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/core.Response"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/core.Response"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/core.Response"
                        }
                    }
                }
            },
            "delete": {
                "description": "根据id删除区块链",
                "tags": [
                    "区块链管理"
                ],
                "summary": "根据id删除区块链",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "主键Id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "{\"code\": 200, message:\"Ok\", \"data\": [...]}",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/core.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/v1.BlockChain"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/core.Response"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/core.Response"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/core.Response"
                        }
                    }
                }
            }
        },
        "/version": {
            "get": {
                "description": "系统版本查询",
                "consumes": [
                    "application/json"
                ],
                "tags": [
                    "系统服务"
                ],
                "summary": "系统版本",
                "responses": {
                    "200": {
                        "description": "{\"code\": 200, \"data\":{\"version\": \"latest\"} }",
                        "schema": {
                            "$ref": "#/definitions/server.Response"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "core.Response": {
            "type": "object",
            "properties": {
                "code": {
                    "description": "数据集",
                    "type": "integer"
                },
                "message": {
                    "type": "string"
                },
                "requestId": {
                    "type": "string"
                }
            }
        },
        "server.Response": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "data": {}
            }
        },
        "v1.BlockChain": {
            "type": "object",
            "properties": {
                "chain_id": {
                    "type": "integer"
                },
                "chain_name": {
                    "type": "string"
                },
                "createdAt": {
                    "type": "string"
                },
                "erc_20_tokens": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/v1.Contract"
                    }
                },
                "extend": {
                    "$ref": "#/definitions/v1.Extend"
                },
                "id": {
                    "type": "integer"
                },
                "infra_http": {
                    "type": "string"
                },
                "infra_websocket": {
                    "type": "string"
                },
                "updatedAt": {
                    "type": "string"
                }
            }
        },
        "v1.BlockChainList": {
            "type": "object",
            "properties": {
                "current_page": {
                    "type": "integer"
                },
                "items": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/v1.BlockChain"
                    }
                },
                "total_count": {
                    "type": "integer"
                },
                "total_page": {
                    "type": "integer"
                }
            }
        },
        "v1.Contract": {
            "type": "object",
            "properties": {
                "chain_id": {
                    "description": "外键",
                    "type": "integer"
                },
                "contract_addr": {
                    "type": "string"
                },
                "decimals": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                },
                "symbol": {
                    "type": "string"
                },
                "type": {
                    "type": "string"
                }
            }
        },
        "v1.Extend": {
            "type": "object",
            "additionalProperties": true
        }
    }
}